cloud_adapter/script/deploy/model_caption/requirements.txt
Flask
cryptography
Pillow